Get daily climate data

Extract daily climate data for a given set of points or polygons within Europe.

get_daily_climate(
  coords = NULL,
  climatic_var = "Prcp",
  period = NULL,
  output = "df"
)

Arguments

coords
A matrix, data.frame, sf::sf(), or terra::SpatVector() object containing point or polygon coordinates in decimal degrees (lonlat/geographic format). Longitude must fall between -40.5 and 75.5 degrees, and latitude between 25.5 and 75.5 degrees. If coords is a matrix, it must have only two columns: the first with longitude and the second with latitude data. If coords is a data.frame, it must contain at least two columns called lon and lat with longitude and latitude coordinates, respectively.

climatic_var
Character. Climatic variable to be downloaded. One of 'Tmax', 'Tmin' or 'Prcp'.

period
Either numbers (representing years between 1950 and 2017), or dates in "YYYY-MM-DD" format (to obtain data for specific days). To specify a sequence of years or dates use the format 'start:end' (e.g. YYYY:YYYY or "YYYY-MM-DD:YYYY-MM-DD", see examples). Various elements can be concatenated in the vector (e.g. c(2000:2005, 2010:2015, 2017), c("2000-01-01:2000-01-15", "2000-02-01"))

output
Character. Either "df", which returns a dataframe with daily climatic values for each point/polygon, or "raster", which returns a terra::SpatRaster() object.

Value

A data.frame or a terra::SpatRaster() object (if output = "raster"). Note temperature values are in ºC\*100, and precipitation values in mm\*100, to avoid floating values.

References

Werner Rammer, Christoph Pucher, Mathias Neumann. 2018. Description, Evaluation and Validation of Downscaled Daily Climate Data Version 2. ftp://palantir.boku.ac.at/Public/ClimateData/

Adam Moreno, Hubert Hasenauer. 2016. Spatial downscaling of European climate data. International Journal of Climatology 36: 1444–1458.

Author

Francisco Rodriguez-Sanchez, Veronica Cruz-Alonso, Sophia Ratcliffe

Examples

if (FALSE) {
library(easyclimate)

# Coords as matrix
coords <- matrix(c(-5.36, 37.40), ncol = 2)
ex <- get_daily_climate(coords, period = "2001-01-01")  # single day
ex <- get_daily_climate(coords, period = c("2001-01-01", "2001-01-03"))  # 1st AND 3rd Jan 2001
ex <- get_daily_climate(coords, period = "2001-01-01:2001-01-03")  # 1st TO 3rd Jan 2001
ex <- get_daily_climate(coords, period = 2008)  # entire year
ex <- get_daily_climate(coords, period = c(2008, 2010))  # 2008 AND 2010
ex <- get_daily_climate(coords, period = 2008:2010)  # 2008 TO 2010

ex <- get_daily_climate(coords, period = "2001-01-01", climatic_var = "Tmin")

# Coords as data.frame
coords <- as.data.frame(coords)
names(coords) <- c("lon", "lat")  # must have these columns
ex <- get_daily_climate(coords, period = "2001-01-01")  # single day

# Coords as sf
coords <- sf::st_as_sf(coords, coords = c("lon", "lat"))
ex <- get_daily_climate(coords, period = "2001-01-01")  # single day

# Several points
coords <- matrix(c(-5.36, 37.40, -4.05, 38.10), ncol = 2, byrow = TRUE)
ex <- get_daily_climate(coords, period = "2001-01-01", output = "raster")  # raster output

## Polygons
coords <- terra::vect("POLYGON ((-5 38, -5 37.5, -4.5 37.5, -4.5 38, -5 38))")

# Return raster
ex <- get_daily_climate(coords, period = "2001-01-01", output = "raster")

# Return dataframe for polygon
ex <- get_daily_climate(coords, period = "2001-01-01")

}